[FLINK-23060] Move FutureUtils#toJava to separate class
